server/requirements.txt
flask==2.2.2
openai==0.27.0
Werkzeug==2.2.2
python-dotenv
googlesearch-python
openai-whisper
gtts
